impl fullfact & pbdesign over common base class, split into new module
